requirements.txt
trio_websocket==0.9.2
trio==0.19.0
requests==2.30
asyncclick==8.1.3.4
pytest_asyncio==0.21.0
